Refactor drive mode handling and add manual mode to config
Moved get_drive_mode function to utils for better modularity and refactored its usage in makemkv.py. Introduced a new column 'manual_mode' in the Config table to track manual mode settings for drives and updated relevant files to accommodate this change. Added a migration script for database schema updates.
