+++ arm/models/config.py
@@ -69,6 +69,7 @@ class Config(db.Model):
     PO_USER_KEY = db.Column(db.String(64))
     PO_APP_KEY = db.Column(db.String(64))
     OMDB_API_KEY = db.Column(db.String(64))
+    manual_mode = db.Column(db.Boolean)
 
     def __init__(self, c, job_id):
         self.__dict__.update(c)
